+++ test/setup.ts
@@ -1,5 +1,3 @@
-import 'vitest-canvas-mock';
-import 'chart.js/auto';
 import '@testing-library/jest-dom/vitest';
 import { cleanup } from '@testing-library/react';
 import axe from 'axe-core';
@@ -21,6 +19,8 @@ afterEach(() => {
   vi.clearAllMocks();
   cleanup();
 });
+
+HTMLCanvasElement.prototype.getContext = (() => {}) as any;
 
 (global as any).ResizeObserver = ResizeObserver;
 (global as any).scrollTo = () => {};