+++ test/visits/charts/DoughnutChart.test.tsx
@@ -1,22 +1,18 @@
 import { screen } from '@testing-library/react';
 import { DoughnutChart } from '../../../src/visits/charts/DoughnutChart';
 import { checkAccessibility } from '../../__helpers__/accessibility';
-import { setUpCanvas } from '../../__helpers__/setUpTest';
+import { renderWithEvents } from '../../__helpers__/setUpTest';
 
 describe('<DoughnutChart />', () => {
-  const stats = {
-    foo: 123,
-    bar: 456,
-  };
-  const setUp = () => setUpCanvas(<DoughnutChart label="This is a chart" stats={stats} />);
+  const stats = { foo: 123, bar: 456 };
+  const dimensions = { width: 800, height: 300 };
+  const setUp = () => renderWithEvents(<DoughnutChart stats={stats} dimensions={dimensions} showNumbersInLegend />);
 
   it('passes a11y checks', () => checkAccessibility(setUp()));
 
   it('renders Doughnut with expected props', () => {
-    const { events } = setUp();
-
-    expect(events).toBeTruthy();
-    expect(events).toMatchSnapshot();
+    const { container } = setUp();
+    expect(container).toMatchSnapshot();
   });
 
   it('renders expected legend', () => {